# docs/openapi.yaml
basePath: /api/v1
definitions:
...
  model.Forum:
    properties:
      category:
        $ref: '#/definitions/model.ForumCategory'
      category_id:
        type: integer
      content:
        type: string
      created_at:
        type: string
      flagged_reason:
        type: string
      has_accepted_answer:
        description: Best answer tracking
        type: boolean
      id:
        type: integer
      is_flagged:
        type: boolean
      is_liked:
        type: boolean
      likes:
        items:
          $ref: '#/definitions/model.ForumLike'
        type: array
      likes_count:
        type: integer
      posts:
        items:
          $ref: '#/definitions/model.ForumPost'
        type: array
      replies_count:
        type: integer
      slug:
        type: string
      title:
        type: string
      trigger_warnings:
        description: Moderation fields
        items:
          type: string
        type: array
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/model.User'
        description: Relations
      user_id:
        type: integer
    type: object
  model.ForumCategory:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
      updated_at:
        type: string
    type: object
  model.ForumLike:
    properties:
      created_at:
        type: string
      forum_id:
        type: integer
      id:
        type: integer
      user_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@ruangtenang.id
    name: API Support
  description: API untuk aplikasi Ruang Tenang - Platform Kesehatan Mental
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Ruang Tenang API
  version: "1.0"
paths:
...
  /forums:
    get:
      consumes:
      - application/json
      description: Get list of forums with pagination and filters
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      - description: Category ID
        in: query
        name: category_id
        type: integer
      - description: 'Support circle: tekanan_akademik, relasi_pertemanan, regulasi_emosi,
          pemulihan_burnout'
        in: query
        name: circle
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get list of forums
      tags:
      - forum
    post:
      consumes:
      - application/json
      description: Create a new forum topic
      parameters:
      - description: Forum request
        in: body
        name: request
        required: true
        schema:
          properties:
            category_id:
              type: integer
            content:
              type: string
            title:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new forum topic
      tags:
      - forum
  /forums/{slug}:
    delete:
      consumes:
      - application/json
      description: Delete a forum (Owner or Admin only)
      parameters:
      - description: Forum Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a forum
      tags:
      - forum
    get:
      consumes:
      - application/json
      description: Get details of a specific forum
      parameters:
      - description: Forum Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Forum'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get forum details
      tags:
      - forum
    post:
      consumes:
      - application/json
      description: Create a reply to a forum topic
      parameters:
      - description: Forum Slug
        in: path
        name: slug
        required: true
        type: string
      - description: Post request
        in: body
        name: request
        required: true
        schema:
          properties:
            content:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a forum post (reply)
      tags:
      - forum
  /forums/{slug}/accepted-answer:
    delete:
      consumes:
      - application/json
      description: Remove the accepted answer mark from a forum (OP only)
      parameters:
      - description: Forum Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Unmark accepted answer for a forum
      tags:
      - forum
    get:
      consumes:
      - application/json
      description: Get the accepted answer for a forum topic
      parameters:
      - description: Forum Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ForumPost'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get accepted answer for a forum
      tags:
      - forum
  /forums/{slug}/like:
    put:
      consumes:
      - application/json
      description: Like or unlike a forum
      parameters:
      - description: Forum Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Toggle forum like
      tags:
      - forum
  /forums/{slug}/posts:
    get:
      consumes:
      - application/json
      description: Get replies for a forum topic with sorting options
      parameters:
      - description: Forum Slug
        in: path
        name: slug
        required: true
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - default: top
        description: 'Sort by: top, newest, oldest'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get forum posts
      tags:
      - forum
...
  /moderation/post-reports:
    get:
      consumes:
      - application/json
      description: Get all pending post reports for moderation (Admin only)
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get pending post reports
      tags:
      - moderation
  /moderation/post-reports/{id}:
    put:
      consumes:
      - application/json
      description: Review and action a post report (Admin only)
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: integer
      - description: Review request
        in: body
        name: request
        required: true
        schema:
          properties:
            notes:
              type: string
            status:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Review a post report
      tags:
      - moderation
...
  /posts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a forum post (Owner or Admin only)
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a forum post
      tags:
      - forum
  /posts/{id}/accept:
    put:
      consumes:
      - application/json
      description: Mark a forum post as the accepted answer (OP only)
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mark post as accepted answer
      tags:
      - forum
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"